[RB] Adds tests for size
